feat(scanning)!: refactors and tweaks, less dependence on `poggers`
Various tweaks and refactors, generally culminating in less reliance on
`poggers`. We now only use it for finding a process by name. I plan on
entirely removing it soon in favor of just using the bare Windows APIs
that `poggers` relies on under the hood (with `windows`). So, for a very
brief time before I add warframe.market integration, this will have only
a dependency on `windows`.