chore: dependency bumps
Bumps versions of dependencies. This also includes several breaking
changes for `windows`, but they are too numerous and too poorly
documented for me to care to check if they are relevant. It appears to
compile and get as far as figuring out that I haven't got Warframe
running, so I'm just hoping that it's fine.